refactor: Diary DTO 패키지
