+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/DiaryController.java
@@ -1,11 +1,11 @@
 package com.startingblue.fourtooncookie.diary;
 
-import com.startingblue.fourtooncookie.diary.dto.request.DiaryFavoriteRequest;
-import com.startingblue.fourtooncookie.diary.dto.request.DiarySaveRequest;
-import com.startingblue.fourtooncookie.diary.dto.request.DiaryUpdateRequest;
-import com.startingblue.fourtooncookie.diary.dto.response.DiaryCreatedResponse;
-import com.startingblue.fourtooncookie.diary.dto.response.DiarySavedResponse;
-import com.startingblue.fourtooncookie.diary.dto.response.DiarySavedResponses;
+import com.startingblue.fourtooncookie.diary.dto.DiaryFavoriteRequest;
+import com.startingblue.fourtooncookie.diary.dto.DiarySaveRequest;
+import com.startingblue.fourtooncookie.diary.dto.DiaryUpdateRequest;
+import com.startingblue.fourtooncookie.diary.dto.DiaryCreatedResponse;
+import com.startingblue.fourtooncookie.diary.dto.DiarySavedResponse;
+import com.startingblue.fourtooncookie.diary.dto.DiarySavedResponses;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
