+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/DiaryService.java
@@ -5,8 +5,8 @@ import com.startingblue.fourtooncookie.character.CharacterService;
 import com.startingblue.fourtooncookie.diary.domain.Diary;
 import com.startingblue.fourtooncookie.diary.domain.DiaryPaintingImageGenerationStatus;
 import com.startingblue.fourtooncookie.diary.domain.DiaryStatus;
-import com.startingblue.fourtooncookie.diary.dto.request.DiarySaveRequest;
-import com.startingblue.fourtooncookie.diary.dto.request.DiaryUpdateRequest;
+import com.startingblue.fourtooncookie.diary.dto.DiarySaveRequest;
+import com.startingblue.fourtooncookie.diary.dto.DiaryUpdateRequest;
 import com.startingblue.fourtooncookie.diary.exception.DiaryDuplicateException;
 import com.startingblue.fourtooncookie.diary.exception.DiaryNotFoundException;
 import com.startingblue.fourtooncookie.diary.service.DiaryLambdaService;
