+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/dto/DiaryCreatedResponse.java
@@ -1,0 +1,4 @@
+package com.startingblue.fourtooncookie.diary.dto;
+
+public record DiaryCreatedResponse(Long diaryId) {
+}
