+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/dto/DiaryFavoriteRequest.java
@@ -1,4 +1,4 @@
-package com.startingblue.fourtooncookie.diary.dto.request;
+package com.startingblue.fourtooncookie.diary.dto;
 
 
 public record DiaryFavoriteRequest(boolean isFavorite) {
