--- fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/dto/response/DiaryCreatedResponse.java
@@ -1,4 +1,0 @@
-package com.startingblue.fourtooncookie.diary.dto.response;
-
-public record DiaryCreatedResponse(Long diaryId) {
-}
